renamed command.datmo for python 2.7
